# openapi.yaml
openapi: 3.0.3
info:
  title: Codex Usage Dashboard API
  version: 1.0.0
  description: |
    当前项目的本地管理 API。服务端使用配置的 ChatGPT OAuth Bearer Token
    读取 wham/usage 和统计接口，前端或其他客户端只需要调用本项目的
    /api/* 接口，不需要直接接触 OAuth 凭证。

    这不是 OpenAI 官方 API，也不是聊天请求代理；它只提供额度、统计、预测、
    配置和提示音资源。

    所有 HTTP 响应都会通过 X-Codex-Meter-Version、X-Codex-Meter-Commit
    和 X-Codex-Meter-Build-Time 响应头提供当前构建标识。
servers:
  - url: http://127.0.0.1:8123
    description: 本地服务示例
  - url: https://codex.your-domain.invalid/codex
    description: 通过反向代理发布时的示例地址，请替换域名和路径
...
paths:
...
  /api/accounts/usage:
    get:
      tags: [accounts]
      operationId: getAccountsUsage
      summary: 获取所有账号额度
      description: 最多并发查询 4 个账号，返回各自额度、缓存状态和错误。单个账号失败不会影响其他账号，整体仍返回 200；存在最近快照时返回 usage 并标记 stale=true 与 error。后台每 5 分钟分别采样所有已配置账号。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      parameters:
        - $ref: '#/components/parameters/Force'
      responses:
        '200':
          description: 多账号额度与逐账号状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsUsageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/accounts/{account_id}:
    parameters:
      - name: account_id
        in: path
        required: true
        description: 本项目账号 id，从账号列表中获取。
        schema:
          type: string
    put:
      tags: [accounts]
      operationId: updateAccount
      summary: 更新账号名称或凭证
      description: 部分更新指定账号，未提供字段保持不变。修改凭证仅使该账号缓存失效；chatgpt_account_id 改变时须新增账号以保留独立历史。代理及其认证属于指定账号；修改代理只使该账号缓存失效。缓存时间和管理认证为全局设置。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: 已更新，返回完整脱敏列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [accounts]
      operationId: deleteAccount
      summary: 删除监控账号
      description: 移除账号凭证和运行时缓存，保留数据库历史归档。删除当前账号后默认切换到列表中第一个账号；禁止删除主账号或最后一个账号，返回 409。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      responses:
        '200':
          description: 已删除，返回剩余脱敏列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AccountConflict'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 服务端启用 Basic Auth 时使用。
    appApiKey:
      type: apiKey
      in: header
      name: X-App-API-Key
      description: 配置 app_api_key 后使用的管理密钥。
    appBearer:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: "与 X-App-API-Key 等价的 Authorization: Bearer 方式。"
  parameters:
...
    Force:
      name: force
      in: query
      required: false
      description: 是否绕过本地缓存直接请求上游。
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: 请求参数或配置无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: force must be true or false
    Unauthorized:
      description: 未通过认证或未提供管理密钥
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
...
    AccountConflict:
      description: 显式指定的账号凭证未配置完整，或尝试删除主账号、最后一个账号
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在或 kind 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Window:
      type: object
      required: [used_percent, reset_at, remaining_seconds]
      properties:
        used_percent:
          type: number
          format: double
          description: 当前窗口已使用百分比，0 到 100。
        window_minutes:
          type: integer
          description: 窗口长度（分钟）。
        reset_after_seconds:
          type: integer
          description: 上游返回的距离重置秒数。
        reset_at:
          type: string
          format: date-time
        remaining_seconds:
          type: integer
          format: int64
          description: 服务端计算的距离重置剩余秒数。
    Credits:
      type: object
      properties:
        has_credits:
          type: boolean
        unlimited:
          type: boolean
        overage_limit_reached:
          type: boolean
        balance:
          description: 上游返回的余额值，通常是字符串。
        approx_local_messages:
          type: array
          items: {type: integer}
        approx_cloud_messages:
          type: array
          items: {type: integer}
    SpendControl:
      type: object
      properties:
        reached:
          type: boolean
        individual_limit: {}
    ResetCredits:
      type: object
      properties:
        available_count:
          type: integer
        applicable_available_count:
          type: integer
    HistoryPoint:
      type: object
      required: [at, used_percent]
      properties:
        at:
          type: string
          format: date-time
        used_percent:
          type: number
          format: double
          description: 采样时本周窗口已使用百分比。
        five_hour_used_percent:
          type: number
          format: double
          nullable: true
          description: 采样时 5 小时窗口已使用百分比；旧日志记录可能没有该字段。
        stale:
          type: boolean
          description: 上游采集异常时是否沿用了最近一次成功采集的数据。
    UsageResponse:
      type: object
      required: [source, rate_limit_allowed, rate_limit_reached, fetched_at, from_cache]
      properties:
        account_id:
          type: string
          description: 数据所属的本项目账号 id，仅在显式指定 account_id 时返回；旧调用省略以保持响应格式。
        account_name:
          type: string
        source:
          type: string
          example: wham_usage
        plan_type:
          type: string
          example: plus
        email:
          type: string
          format: email
        rate_limit_allowed:
          type: boolean
        rate_limit_reached:
          type: boolean
        rate_limit_reached_type:
          type: string
        credits:
          allOf:
            - $ref: '#/components/schemas/Credits'
        spend_control:
          allOf:
            - $ref: '#/components/schemas/SpendControl'
        rate_limit_reset_credits:
          allOf:
            - $ref: '#/components/schemas/ResetCredits'
        fetched_at:
          type: string
          format: date-time
        from_cache:
          type: boolean
        five_hour:
          allOf:
            - $ref: '#/components/schemas/Window'
        seven_day:
          allOf:
            - $ref: '#/components/schemas/Window'
        history:
          type: array
          description: 兼容字段，返回周额度和 5 小时额度独立时间线的合并结果，按采样时间升序排列；合并结果最多包含 96 条记录。新客户端应使用 weekly_history 和 five_hour_history。
          items:
            $ref: '#/components/schemas/HistoryPoint'
        weekly_history:
          type: array
          description: 从完整定时采样归档计算的周额度独立时间线，先按自身 used_percent 去重并优先保留每个百分比最近的正常采样（没有正常采样时保留最近降级采样），再按采样时间升序返回最近最多 48 个不同百分比；不受 5 小时额度变化影响。
          items:
            $ref: '#/components/schemas/HistoryPoint'
        five_hour_history:
          type: array
          description: 从完整定时采样归档计算的 5 小时额度独立时间线，先按自身 five_hour_used_percent 去重并优先保留每个百分比最近的正常采样（没有正常采样时保留最近降级采样），再按采样时间升序返回最近最多 48 个不同百分比；不受周额度变化影响。
          items:
            $ref: '#/components/schemas/HistoryPoint'
...
    AccountView:
      type: object
      required: [id, name, active, primary, token_configured, cookie_configured, setup_required, proxy_password_configured]
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
        primary:
          type: boolean
          description: 是否为兼容旧应用 API 的固定主账号，主账号不能通过账号删除接口移除。
        token_configured:
          type: boolean
        token_hint:
          type: string
          description: Token 的脱敏提示，仅包含末四位。
        cookie_configured:
          type: boolean
        chatgpt_account_id:
          type: string
        user_agent:
          type: string
        fedramp:
          type: boolean
        client_build_number:
          type: string
        client_version:
          type: string
        device_id:
          type: string
        session_id:
          type: string
        client_observation:
          type: string
        referer:
          type: string
        setup_required:
          type: boolean
        proxy_url:
          type: string
          description: 该账号独立的代理 URL，密码以 **** 脱敏；为空或省略表示直连。

        proxy_password_configured:
          type: boolean
          description: 该账号是否已保存代理密码，不返回密码原文。
    AccountsView:
      type: object
      required: [active_account_id, primary_account_id, accounts]
      properties:
        active_account_id:
          type: string
          description: 页面持久化展示的账号，独立于旧应用 API 默认账号。
        primary_account_id:
          type: string
          description: 旧应用 API 固定使用的主账号；优先 id=default，否则为账号列表第一项。
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountView'
    AccountUpdate:
      type: object
      description: 新增时必须提供 access_token 和 chatgpt_account_id；更新时未提供字段保持不变。仅接收账号相关字段。
      properties:
        name:
          type: string
          maxLength: 80
        access_token:
          type: string
          writeOnly: true
        cookie:
          type: string
          writeOnly: true
        chatgpt_account_id:
          type: string
        user_agent:
          type: string
        fedramp:
          type: boolean
        client_build_number:
          type: string
        client_version:
          type: string
        device_id:
          type: string
        session_id:
          type: string
        client_observation:
          type: string
        referer:
          type: string
        proxy_url:
          type: string
          description: 该账号的代理 URL，支持 http、https、socks5；空字符串明确选择直连。省略保持不变，地址不带认证时保留已有认证。
        proxy_username:
          type: string
          writeOnly: true
          description: 代理认证用户名。更改用户名时应同时提供新密码。
        proxy_password:
          type: string
          writeOnly: true
          description: 代理认证密码，支持特殊字符，无需 URL 编码。省略保留已有密码，空字符串明确设置空密码；响应不回显。
        proxy_clear_auth:
          type: boolean
          default: false
          description: 为 true 时清除该账号的代理认证，不能同时提供非空用户名或密码。
...
    AccountUsage:
      type: object
      required: [id, name, active, primary, stale]
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
        primary:
          type: boolean
          description: 是否为兼容旧应用 API 的固定主账号，主账号不能通过账号删除接口移除。
        stale:
          type: boolean
          description: 上游失败且 usage 是最近缓存快照时为 true。
        error:
          type: string
        usage:
          $ref: '#/components/schemas/UsageResponse'
    AccountsUsageResponse:
      type: object
      required: [active_account_id, primary_account_id, accounts]
      properties:
        active_account_id:
          type: string
          description: 页面持久化展示的账号，独立于旧应用 API 默认账号。
        primary_account_id:
          type: string
          description: 旧应用 API 固定使用的主账号；优先 id=default，否则为账号列表第一项。
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountUsage'
...